Update consent-form.md
Note that I made all of these changes also in qualtrics. So if you do not like them, please reverse the changes in qualtrics, too.

I changed the discretion sentence as discretion also means that everyone remains quiet about it, but we will ask them for the time during the interview.

In addition, why does it say that it takes at least 30 min?