Update instructions.md
Again, all changes made here are also made in qualtrics.

Maybe, we want to the rename the spreadsheet?

The download of the instructions does not work. Or more precisely: I can download it but cannot open it with Adobe.
Further, I would rename the file so that nothing with GitHub shows up in it (because otherwise they may search for this site here).

Finally, the download of the spreadsheets works but the formula in the cells for it is wrong.

If I click on the doodle link, doodle is opened in the same tab (so I am no longer in qualtrics).

The references that you made in text (e.g., schedule here worked sometimes, but not always). Therefore, I have them replaced in qualtrics but not here